requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.10
django-bootstrap-form==3.4
django-filter==21.1
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
importlib-metadata==4.10.1
Markdown==3.3.6
Pillow==9.0.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.5
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
webencodings==0.5.1
xhtml2pdf==0.2.5
xlwt==1.3.0
zipp==3.7.0
